AB_Testing
An unnamed company first introduced a new type of bid, "average bidding", as an alternative to the current bidding type called "maximum bidding".
A customer of ours has decided to test this new feature and wants to do an A/B test to see if average bidding converts more than maximum bidding.